refactor: seperate info-form
